requirements.txt
anthropic
streamlit
plotly
feedparser
beautifulsoup4
requests
python-dotenv
apscheduler
yfinance
reportlab
playwright
